+++ baybe/recommenders/pure/bayesian/botorch.py
@@ -10,7 +10,7 @@ from typing import TYPE_CHECKING, Any, ClassVar
 
 import numpy as np
 import pandas as pd
-from attrs import define, field
+from attrs import define, field, fields
 from attrs.converters import optional as optional_c
 from attrs.validators import ge, gt, instance_of
 from typing_extensions import override
@@ -348,6 +348,17 @@ class BotorchRecommender(BayesianRecommender):
             if isinstance(p, _FixedNumericalContinuousParameter)
         }
 
+        if (
+            self.sequential_continuous
+            and subspace_continuous.has_interpoint_constraints
+        ):
+            raise IncompatibilityError(
+                f"The '{fields(BotorchRecommender).sequential_continuous.name}' flag "
+                f"is set to ``True``, but interpoint constraints are present in the "
+                f"continuous subspace. This is not supported. "
+                f"Please set it to ``False``."
+            )
+
         # NOTE: The explicit `or None` conversion is added as an additional safety net
         #   because it is unclear if the corresponding presence checks for these
         #   arguments is correctly implemented in all invoked BoTorch subroutines.
@@ -360,12 +371,18 @@ class BotorchRecommender(BayesianRecommender):
             raw_samples=self.n_raw_samples,
             fixed_features=fixed_parameters or None,
             equality_constraints=[
-                c.to_botorch(subspace_continuous.parameters)
+                c.to_botorch(
+                    subspace_continuous.parameters,
+                    batch_size=batch_size if c.is_interpoint else None,
+                )
                 for c in subspace_continuous.constraints_lin_eq
             ]
             or None,
             inequality_constraints=[
-                c.to_botorch(subspace_continuous.parameters)
+                c.to_botorch(
+                    subspace_continuous.parameters,
+                    batch_size=batch_size if c.is_interpoint else None,
+                )
                 for c in subspace_continuous.constraints_lin_ineq
             ]
             or None,
@@ -407,6 +424,13 @@ class BotorchRecommender(BayesianRecommender):
             The recommended points.
         """
         assert self._objective is not None
+
+        # TODO Defining interpoint constraints for hybrid spaces is probably not
+        # that hard, but should be investigated in a follow up PR.
+        if searchspace.continuous.has_interpoint_constraints:
+            raise IncompatibilityError(
+                "Interpoint constraints are not available in hybrid spaces."
+            )
         if (
             batch_size > 1
             and not self._get_acquisition_function(self._objective).supports_batching
@@ -453,6 +477,7 @@ class BotorchRecommender(BayesianRecommender):
                 c.to_botorch(
                     searchspace.continuous.parameters,
                     idx_offset=len(candidates_comp.columns),
+                    batch_size=batch_size if c.is_interpoint else None,
                 )
                 for c in searchspace.continuous.constraints_lin_eq
             ]
@@ -461,6 +486,7 @@ class BotorchRecommender(BayesianRecommender):
                 c.to_botorch(
                     searchspace.continuous.parameters,
                     idx_offset=num_comp_columns,
+                    batch_size=batch_size if c.is_interpoint else None,
                 )
                 for c in searchspace.continuous.constraints_lin_ineq
             ]